Implement GUI partially
