+++ src/main/java/rua/common/Ui.java
@@ -1,10 +1,30 @@
 package rua.common;
 
+import javafx.application.Application;
+import javafx.scene.layout.Region;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.control.Label;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.VBox;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
+import javafx.stage.Stage;
 import java.io.InputStream;
 import java.io.PrintStream;
 import java.util.Scanner;
 
-public class Ui {
+public class Ui extends Application {
+    private ScrollPane scrollPane;
+    private VBox dialogContainer;
+    private TextField userInput;
+    private Button sendButton;
+    private Scene scene;
+    private Image user = new Image(this.getClass().getResourceAsStream("/images/DaUser.png"));
+    private Image duke = new Image(this.getClass().getResourceAsStream("/images/DaRua.png"));
+
     private final Scanner in;
     private final PrintStream out;
 
@@ -17,36 +37,142 @@ public class Ui {
         this.out = out;
     }
 
-    public void showLine() {
-        out.println("____________________________________________________________");
+    public String showLine() {
+        return "____________________________________________________________";
     }
 
     public void showWelcome() {
-        String greeting = "____________________________________________________________\n"
+        String welcome = "____________________________________________________________\n"
                 + " Hello! I'm Rua, your ChatBot\n"
                 + " What can I do for you?\n"
                 + "____________________________________________________________";
-        out.println(greeting);
+        dialogContainer.getChildren().add(
+                DialogBox.getDukeDialog(new Label(welcome), new ImageView(duke))
+        );
     }
 
-    public void showGoodbye() {
-        String goodbye = " Bye. Hope to see you again soon!";
-        out.println(goodbye);
+    public String showGoodbye() {
+        return " Bye. Hope to see you again soon!";
     }
 
-    public void showMessage(String str) {
-        out.println(str);
+    public String showMessage(String str) {
+        return str;
     }
 
-    public void showError(String errorMessage) {
-        out.println("You get an error: " + errorMessage);
+    public String showError(String errorMessage) {
+        return "You get an error: " + errorMessage;
     }
 
-    public void showLoadingError() {
-        showError("Given tasklist cannot be loaded. Now creating a new tasklist instead.");
+    public String showLoadingError() {
+        return showError("Given tasklist cannot be loaded. Now creating a new tasklist instead.");
     }
 
     public String readCommand() {
-        return in.nextLine();
+        return userInput.getText();
+    }
+
+    @Override
+    public void start(Stage stage) {
+        // Step 1 code here
+        //The container for the content of the chat to scroll.
+        scrollPane = new ScrollPane();
+        dialogContainer = new VBox();
+        scrollPane.setContent(dialogContainer);
+
+        userInput = new TextField();
+        sendButton = new Button("Send");
+
+        AnchorPane mainLayout = new AnchorPane();
+        mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
+
+        scene = new Scene(mainLayout);
+
+        stage.setScene(scene);
+        stage.show();
+
+        //Step 2 code here
+        stage.setTitle("Rua");
+        stage.setResizable(false);
+        stage.setMinHeight(600.0);
+        stage.setMinWidth(500.0);
+
+        mainLayout.setPrefSize(500.0, 600.0);
+
+        scrollPane.setPrefSize(485, 535);
+        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
+        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
+
+        scrollPane.setVvalue(1.0);
+        scrollPane.setFitToWidth(true);
+
+        // You will need to import `javafx.scene.layout.Region` for this.
+        dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
+
+        userInput.setPrefWidth(425.0);
+
+        sendButton.setPrefWidth(55.0);
+
+        AnchorPane.setTopAnchor(scrollPane, 1.0);
+
+        AnchorPane.setBottomAnchor(sendButton, 1.0);
+        AnchorPane.setRightAnchor(sendButton, 1.0);
+
+        AnchorPane.setLeftAnchor(userInput , 1.0);
+        AnchorPane.setBottomAnchor(userInput, 1.0);
+
+        //Step 3. Add functionality to handle user input.
+        sendButton.setOnMouseClicked((event) -> {
+            dialogContainer.getChildren().add(getDialogLabel(userInput.getText()));
+            userInput.clear();
+        });
+
+        userInput.setOnAction((event) -> {
+            dialogContainer.getChildren().add(getDialogLabel(userInput.getText()));
+            userInput.clear();
+        });
+        dialogContainer.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
+
+        //showWelcome();
+        //Part 3. Add functionality to handle user input.
+        sendButton.setOnMouseClicked((event) -> {
+            handleUserInput(userInput.getText());
+        });
+
+        userInput.setOnAction((event) -> {
+            handleUserInput(userInput.getText());
+        });
+    }
+
+    /**
+     * Iteration 1:
+     * Creates a label with the specified text and adds it to the dialog container.
+     * @param text String containing text to add
+     * @return a label with the specified text that has word wrap enabled.
+     */
+    private Label getDialogLabel(String text) {
+        // You will need to import `javafx.scene.control.Label`.
+        Label textToAdd = new Label(text);
+        textToAdd.setWrapText(true);
+
+        return textToAdd;
+    }
+
+
+    public void handleUserInput(String text) {
+        Label userText = new Label(text);
+        Label dukeText = new Label(getResponse(text));
+        dialogContainer.getChildren().addAll(
+                DialogBox.getUserDialog(userText, new ImageView(user)),
+                DialogBox.getDukeDialog(dukeText, new ImageView(duke))
+        );
+        userInput.clear();
+    }
+
+    /**
+     * You should have your own function to generate a response to user input.
+     * Replace this stub with your completed method.
+     */
+    private String getResponse(String input) {
+        return "Rua heard: " + input;
     }
 }
